==> 1_Import_Clean.ipynb <==
# First notebook
## Importing documents and first cleaning.

### Information to consider according to the client:
- Classification model (Different targets), to predict best seller, supplier, product, line product, features, etc. 
- Regression model: Value per period of time
- Cost, profit and profit margin 
- Profit relation (supplier, seller, product, line, group, etc.)
- Inventory Turnover 

## This notebook uses the following libraries: 
- pandas as pd: pandas is a library for data manipulation and analysis. It is imported and given the alias pd.
- glob: glob is a library for finding files and directories matching a specific pattern. It is used for searching for files.
- os: os is a library for interacting with the operating system. It is used for interacting with the file system.
- numpy as np: numpy is a library for numerical computing. It is imported and given the alias np.
- tqdm(tqdm.notebook): A progress bar for Jupyter notebooks

# import libraries
import pandas as pd
import numpy as np
import glob
import os
from tqdm.notebook import tqdm

#### The client´s documents where retrieved from a local ERP with limitation, therefore, I will import all the files and analyze what can be done from the clients requierements

%%time
# import sales documents from 2021 and concatenate them in 1 single database (& check time)
all_files = glob.glob(os.path.join(r"C:\Users\mfduh\OneDrive\Documents\Deutschland\Arbeiten\Data Analyst\Jupyter excercises\Week 9 - Final Project\Company_Documents\Projekt", "*.xls"))
ventas2021 = pd.concat((pd.read_excel(f) for f in tqdm(all_files, desc="Reading files")), ignore_index=True)

%%time
# import sales documents from 2022 and concatenate them in 1 single database (& check time)
all_files = glob.glob(os.path.join(r"C:\Users\mfduh\OneDrive\Documents\Deutschland\Arbeiten\Data Analyst\Jupyter excercises\Week 9 - Final Project\Company_Documents\2022 - Mensual", "*.xls"))
ventas2022 = pd.concat((pd.read_excel(f) for f in tqdm(all_files, desc="Reading files")), ignore_index=True)

%%time
# import documents of the warehouse from 2021 and concatenate them in 1 single database (& check time)
all_files = glob.glob(os.path.join(r"C:\Users\mfduh\OneDrive\Documents\Deutschland\Arbeiten\Data Analyst\Jupyter excercises\Week 9 - Final Project\Company_Documents\2021 Existencias", "*.xls"))
exist2021 = pd.concat((pd.read_excel(f) for f in tqdm(all_files, desc="Reading files")), ignore_index=True)

%%time
# import documents of the warehouse from 2022 and concatenate them in 1 single database (& check time)
all_files = glob.glob(os.path.join(r"C:\Users\mfduh\OneDrive\Documents\Deutschland\Arbeiten\Data Analyst\Jupyter excercises\Week 9 - Final Project\Company_Documents\2022 Existencias", "*.xls"))
exist2022 = pd.concat((pd.read_excel(f) for f in tqdm(all_files, desc="Reading files")), ignore_index=True)

# allow the option to view all the columns of the document
pd.set_option('display.max_columns', None)

# concatenate the sales documents to have them as 1
ventas = pd.concat([ventas2021, ventas2022], axis=0)

# reseting index
ventas = ventas.reset_index(drop=True)

# check the database
ventas.head(3)

# check NaN values and display max_rows to view them correctly
pd.set_option('display.max_rows', None)
# ventas.isna().sum()

# check specific columns
ventas['RANGO_MARGEN'].isna().sum()

# There are many columns with no or very little information, after checking them, I decided to drop columns with a thresh of 600,000
ventas = ventas.dropna(axis = 1, thresh = 600000)

# check NaN values again
ventas.isna().sum()

# checked the different columns, values and relevance for the Analysis 
# and return max_rows to 10
pd.set_option('display.max_rows', 10)
ventas['VENDEDOR'].value_counts(dropna = False)

ventas.head(3)

# After speaking to the client: drop unnecesary columns
# A couple are pending to check the utility, meanwhile they´ll be kept
ventas = ventas.drop(columns=['FI', 'FF', 'TIPO', 'NALMACEN','FECHA', 'FOLIO', 'FACTURA_CORTE', 
                            'TIPOF', 'CLIENTE', 'NOMBRE', 'ZONA', 'TIPO_CLIENTE', 'NVENDEDOR', 
                              'GIRO', 'GRUPO', 'SUBGRUPO', 'LISTAPRECIOS', 'IEPS', 'UTILIDAD',
                              'MARGEN', 'UNIDAD', 'RANGO_MARGEN', 'NPROV1', 'TASA_IEPS',
                              'TASA_IVA','USR5', 'PVCORTE', 'FECHA_PVCORTE', 'PVA', 'EXPORTACION',
                              'DESCTO_EN_LISTA_CLIENTE', 'CLAVERANGO', 'CAJA', 'REMISION'], axis=1)

ventas.shape

# standarize columns
def column_change(df):
    cols = []
    for column in range(len(df.columns)):
        cols.append(df.columns[column].lower().replace(' ', '_'))
    df.columns = cols
    return

column_change(ventas)

ventas

# drop last row as it is only NaN values
ventas = ventas.drop(index=1549332)

# check NaN values and display max_rows to view them correctly
pd.set_option('display.max_rows', None)
ventas.isna().sum()

# save the first draft of the document that will be used to analyse as a CSV.
ventas.to_csv('ventas_completo.csv', index=False)

### Check documents of the warehouse

# concatenate both years
existencias = pd.concat([exist2021, exist2022], axis=0)

# reset indexes
existencias = existencias.reset_index(drop=True)

# display max_rows to 10 and check the document
pd.set_option('display.max_rows', 10)
existencias

# drop last row as it is only NaN values
existencias = existencias.drop(index=279943)

existencias.head(2)

existencias.isna().sum()

# save the first draft of the document that will be used to analyse as a CSV.
existencias.to_csv('existencias_completo.csv', index=False)
