==> 2_Clean_SQL.ipynb <==
# Second notebook 
## Importing cleaned CSV, continue cleaning and export to SQL
- Focus: sales

# import libraries
import pandas as pd
import numpy as np
import glob
import os
from tqdm.notebook import tqdm
import mysql.connector
import getpass 
from sqlalchemy import create_engine

# import saved CSV
data = pd.read_csv('ventas_completo.csv')

data.dtypes

pd.set_option('display.max_columns', None)
# pd.set_option('display.max_rows', None)

data = data.dropna(axis = 0, thresh = 20)

#Check values and human errors
# data['talla'].value_counts()
# data['prov1'].value_counts()

# replace specific values that have not been changed due to the human mistakes when filling the ERP.

data.prov1 = data.prov1.replace([37.0,102.0,8.0,210.0,190.0,'#N/A!',1.0,14.0,20.0,27.0,'}0144',165.0,2.0,123.0,
                                   191.0,163.0,179.0, 189.0,4.0,169.0,126.0, 29.0,15.0,183.0,129.0,
                                  168.0,130.0,173.0,182.0,'008',119.0,35.0,28.0,144.0,141.0,3.0,112.0,
                                  76.0,158.0,49.0,194.0,142.0,22.0,32.0,196.0,16.0,178.0,53.0,30.0,
                                  21.0,18.0,33.0], 
                                ['0037','0102','0008', '0210', '0190', 'Other', '0001','0014', '0020', '0027', '0144', '0165', '0002','0123',
                                   '0191','0163', '0179', '0189' , '0004', '0169', '0126', '0029','0015', '0183', '0129',
                                  '0168', '0130', '0173', '0182', '0008', '0119', '0035','0028','0144', '0141', '0003','0112',
                                  '0076', '0158', '0049','0194','0142','0022', '0032', '0196', '0016', '0178', '0053', '0030',
                                  '0021','0018','0033'])

# Fill NaN
data['talla'] = data['talla'].fillna('T- UNITALLA')
data['modelo'] = data['modelo'].fillna('ALGODÓN')
data['color'] = data['color'].fillna('SURTIDO')
data['nsubgrupo'] = data['nsubgrupo'].fillna('LINEA')
data['lista'] = data['lista'].fillna('AGUASCALIENTES')
data['descuento'] = data['descuento'].fillna(0)
data['prov1'] = data['prov1'].fillna('Other')
data['usr1'] = data['usr1'].fillna('Other')

data.isna().sum()

# Checking for NaN in usr1
# pd.set_option('display.max_colwidth', None)
# data.loc[220:250]
# data[data['usr1'].isna()==True].head()

# rename column names to more descriptive ones for easier identification 

data = data.rename(columns={'ngrupo': 'grupo', 'nsubgrupo': 'subgrupo', 'prov1':'proveedor', 'usr1':'detalle'})

# reset index

data = data.reset_index(drop=True)

# view different sets of data to observe better the changes 

data.head(747)

# check the stores/warehouses total amount of appearences from all the products
data['almacen'].value_counts()

# I noted there was an additional "clase" identified as a "cancelación" which generates negative products and messes up the counts.

data[data.eq("NENUCO").any(1)]

### Start working with the warehouse DF

# read saved document
costs = pd.read_csv('existencias_completo.csv')

# check count for 'almacen'
costs['ALMACEN'].value_counts(dropna=False)

# delete the 'almacen' that has no value where the name does not exist
costs = costs[~costs['NALMACEN'].isna()]

# reset index
costs = costs.reset_index(drop=True)

# check the amount of unique products that exist
costs['PRODUCTO'].nunique()
# pd.set_option('display.max_rows', None)

# check duplicates 
costs.duplicated().sum()

# check shape
costs.shape

# drop duplicates 
costs = costs.drop_duplicates()

# standarize columns

def column_change(df):
    cols = []
    for column in range(len(df.columns)):
        cols.append(df.columns[column].lower().replace(' ', '_'))
    df.columns = cols
    return

column_change(costs)

# reset index and sort by product 

costs = costs.reset_index(drop=True)
costs.sort_values(by='producto').head(2)

# Checking the relationship between products and costs, because not all values where given and it is important to fill the price
# and cost for each one, I´m identifying what is actually given in the ERP

costs[costs["producto"] == 35]
# costo_base is not the average of the cost; rather just one random used 
costs[costs["producto"] == 354]
# or no clear visual of how it was calculated; as there are values marked as 0 in costo_unitario;
# I will use costo_base as the cost, but will inform the client about this.

# identify unique products
costs['producto'].nunique()

# check how are they categorized to be able to deal withe them better
costs.dtypes

# due to price change per month, I decided to use the price stablished in december 2021; and for now drop all
# the extra information.
costs = costs.drop_duplicates(subset=['producto'],keep='first').reset_index(drop=True)


# keep only the relevant information in a new df called 'costo'
costo = costs[['producto','costo_base','precio_1','precio_2']]

# check the price_1 and price_2 that was given within the ERP to determine which will be used after speaking with the client
costo.head(3)

data.head(3)
data['producto'].nunique()

# I will drop cost and send my DF to SQL to make queries for models depending on the clients needs.
venta = data.copy()
venta.head(3)

# dropping 'costo' as it is not relevant in this DF and we will fill the information with the other DF
venta = venta.drop(columns=['costo'], axis=1)
venta

# Save new dataframes into CSV for future use. 

venta.to_csv('cleaned_venta.csv', index=False)
costo.to_csv('cleaned_costo.csv', index=False)

data

# create a new dataframe in MySQL
# generate a getpass and connect to Workbench

passw = getpass.getpass()

mydb = mysql.connector.connect(
  host="localhost",
  user="root",
  password=passw
)

# export information to SQL to generate queries for different models.
# name I gave it was: final_project; it is created inside Workbench.
# tables created: 'ventas', 'costo', 'data'

connection_string = 'mysql+pymysql://root:'+passw+'@localhost/final_project'
engine = create_engine(connection_string)

venta.to_sql(name = 'ventas', con=engine, if_exists='replace', index=False)
costo.to_sql(name = 'costo', con=engine, if_exists='replace', index=False)
data.to_sql(name = 'data', con=engine, if_exists='replace', index=False)
