==> 3_SQL_to_csv_for_modeling.ipynb <==
# import libraries
import pandas as pd
import numpy as np
import glob
import os
import mysql.connector
import getpass 
from sqlalchemy import create_engine

import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.datasets import load_boston
from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import train_test_split

from sklearn.model_selection import train_test_split
from sklearn import linear_model
from sklearn.metrics import mean_squared_error, r2_score, mean_absolute_error
import math
from sklearn.preprocessing import MinMaxScaler 
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import OneHotEncoder
from sklearn.ensemble import RandomForestRegressor

# Bring table from MySQL
password = getpass.getpass()

connection_string = 'mysql+pymysql://root:'+password+'@localhost/kaggle_project'
engine = create_engine(connection_string)

sales = pd.read_sql_query('SELECT * FROM sales_stores', engine)

## Generate different Queries to obtain relevant information for the client
### Will save all of them in CSVs and DF here

# cancelations per sales person
cancelaciones_por_vendedor = pd.read_sql_query('''SELECT count(clase) as cancelaciones, vendedor FROM final_project.ventas
WHERE clase = 'CANCELACION'
group by vendedor
order by cancelaciones desc
''', engine)

cancelaciones_por_vendedor.to_csv('cancelaciones_por_vendedor.csv', index=False)

cancelaciones_por_vendedor

# product price, cost and profit
precio_utilidad_por_producto = pd.read_sql_query('''
SELECT v.producto, round(c.precio_1,2) as precio_1, round(c.precio_2,2) as precio_2, round(c.costo_base,2) as costo, 
round((c.precio_1 - c.costo_base),2) as utilidad_1, round((c.precio_2 - c.costo_base),2) as utilidad_2
FROM final_project.ventas v 
JOIN final_project.costo c USING (producto)
WHERE clase = 'VENTA'
group by v.producto
order by producto
''', engine)

precio_utilidad_por_producto.to_csv('precio_utilidad_por_producto.csv', index=False)

precio_utilidad_por_producto

# there are products with loss

# productos vendidos por almacen
producto_vendido_por_almacen_total = pd.read_sql_query('''
SELECT v.almacen, count(v.producto) as productos_vendidos FROM final_project.ventas v
group by almacen
order by productos_vendidos desc
''', engine)

producto_vendido_por_almacen_total.to_csv('producto_vendido_por_almacen_total.csv', index=False)

producto_vendido_por_almacen_total

# productos vendidos por almacen por mes
producto_vendido_por_almacen_mensual = pd.read_sql_query('''
SELECT v.almacen, v.ejercicio, v.mes, count(v.producto) as productos_vendidos FROM final_project.ventas v
group by almacen, ejercicio, mes
order by productos_vendidos desc
''', engine)

producto_vendido_por_almacen_mensual.to_csv('producto_vendido_por_almacen_mensual.csv', index=False)

producto_vendido_por_almacen_mensual

# full table before including NaN costs
costo_para_llenar_nan = pd.read_sql_query('''
SELECT d.*, c.costo_base as 'costo_ind' FROM final_project.data d
JOIN final_project.costo c USING (producto)
WHERE clase = 'VENTA'
''', engine)

costo_para_llenar_nan.to_csv('costo_para_llenar_nan.csv', index=False)

costo_para_llenar_nan
