==> 4_final_DF_to_SQL.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np
from sklearn.datasets import load_boston
from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import train_test_split
import mysql.connector
import getpass 
from sqlalchemy import create_engine
from sklearn.model_selection import train_test_split
from sklearn import linear_model
from sklearn.metrics import mean_squared_error, r2_score, mean_absolute_error
import math
from sklearn.preprocessing import MinMaxScaler 
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import OneHotEncoder
from sklearn.ensemble import RandomForestRegressor

pd.set_option('display.max_columns', None)

# cancelaciones_por_vendedor = pd.read_csv('cancelaciones_por_vendedor.csv')
# precio_utilidad_por_producto = pd.read_csv('precio_utilidad_por_producto.csv')
# producto_vendido_por_almacen_total = pd.read_csv('producto_vendido_por_almacen_total.csv')
# producto_vendido_por_almacen_mensual = pd.read_csv('producto_vendido_por_almacen_mensual.csv')
costo_para_llenar_nan = pd.read_csv('costo_para_llenar_nan.csv', low_memory=False)

costo_para_llenar_nan

costo_para_llenar_nan.isnull().sum()

# as this is ia big df np.where is better than a loop. 
costo_para_llenar_nan['costo'] = np.where(costo_para_llenar_nan['costo'].isna(), 
                                          costo_para_llenar_nan['costo_ind']*costo_para_llenar_nan['cantidad'],
                                          costo_para_llenar_nan['costo'])

costo_para_llenar_nan.isna().sum()

##### this is not the best solution (for big DF), kept it for reference
# for i in range(len(costo_para_llenar_nan)):
#     if costo_para_llenar_nan['costo'].isna().iloc[i] == True:
#         costo_para_llenar_nan[costo_para_llenar_nan['costo'].iloc[i]] = costo_para_llenar_nan['costo_ind'].iloc[i] * costo_para_llenar_nan['cantidad'].iloc[i]
#     else:
#         costo_para_llenar_nan[costo_para_llenar_nan['costo'].iloc[i]] = costo_para_llenar_nan[costo_para_llenar_nan['costo'].iloc[i]]*1

# create a new dataframe in MySQL
passw = getpass.getpass()

mydb = mysql.connector.connect(
  host="localhost",
  user="root",
  password=passw
)

mycursor = mydb.cursor()
mycursor.execute("CREATE DATABASE aura")

# name I gave it was: aura; it is created inside Workbench

# export information to SQL to generate queries for different models.

connection_string = 'mysql+pymysql://root:'+passw+'@localhost/aura'     # here we choose the data base, here: midproject
engine = create_engine(connection_string)

costo_para_llenar_nan.to_sql(name = 'df', con=engine, if_exists='replace', index=False)

costo_para_llenar_nan.to_csv('aura_df.csv', index=False)
